[dagit] Track GanttChart mode in localStorage
Summary:
Track GanttChart view mode in localStorage so that it persists across runs.

Converted `GanttChart` to be an FC with the ref behavior moved into an effect, so it wouldn't hurt to patch this in and poke at it if you have some time, just to sanity check.

Test Plan: View Run in Dagit, verify that the Gantt renders properly. Change view mode, refresh page. Verify that the mode has persisted correctly.

Reviewers: bengotow

Reviewed By: bengotow

Subscribers: jordansanders

Differential Revision: https://dagster.phacility.com/D8048