build_init_logger_context
Summary:
This diff adds build_logger_init_context, in response to a user that requested it https://github.com/dagster-io/dagster/discussions/4174.
There's some weirdness here around having to provide a logger definition to the function. An alternative could be to make this a method on the logger definition itself, but that doesnt fit in with the other build_x_context patterns.

There's also the question of pipeline_def and run id. Should these be exposed at top level?

Test Plan: Added unit tests

Reviewers: alangenfeld, sandyryza, prha

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D8079