Support copying multiple directories
Summary:
https://github.com/dagster-io/dagster/commit/de61b43f4849c965fcee4d1274c9591fcbf753d1
incorrectly nested the copy statements when rearranging the try/finally
blocks. This caused an issue with the 0.11.12 release where we were
unable to build k8s-dagit-example because it was unable to correctly
copy multiple directories.

This corrects the nesting and adds additional test coverage to flex the
"copy multiple directories" behavior (as well as to ensure this breaks
if the directory is not nested inside a git repo).

Test Plan: unit

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D8246