[dagit] Fix stdout/stderr log view
Summary:
Resolves #4254.

When viewing the stdout/stderr option in Run logs, the step selector doesn't set the selected value correctly. Links from the logs themselves also do not navigate to the correct step. Fix this by tracking `logKey` in the URL query. This allows us to link directly to the view of that log key in the log item itself, and repairs the select element.

Test Plan:
View a run. Click to stdout view, change selected step. Verify that the select updates properly, the URL updates properly, and the stdout/stderr is correct.

Click a "View stdout/stderr" link on a log item. Verify that it goes to the correct log key in the stdout view.

Reload the page with query param set, verify that it loads the stdout view correctly.

Reviewers: bengotow, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D8273