Resolve Versions fix
Summary: If the original `io_manager` resource key is not in use, and a user provides an additional key for an io manager, the default `io_manager` key will still have a config entry. This breaks an old invariant in the version resolution code path. Removed this outdated invariant and added a test to ensure correct behavior.

Test Plan: unit

Reviewers: sandyryza, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D8306