Be less strict in dagster image build CLI about the environment matching the passed in version
Summary: When we call this in the release pipeline, it's called from the release branch, so there's no need to be super-strict about the version matching perfectly. Just make sure that the correct dagster version is passed into the docker container to use when making the build.

Test Plan: Run an RC release, should go through the first time

Reviewers: jordansanders, max

Reviewed By: jordansanders

Differential Revision: https://dagster.phacility.com/D8339