Demonstrate using the fixtures in a test suite
Summary:
This shows both how to activate the fixtures (by adding them to the
`pytest_plugins` variable) and how to use them to write a real test.

If we like this approach, I'll eventually cut over our other uses
of docker-compose which will simplify our .buildkite/ files quite a
bit.

I had to modify the docker compose fixtures to conditionally switch
between "docker-compose" when no context is passed and "docker compose"
when a context is passed - I assume this is a rough edge of the new
docker compose cli? On Docker for Mac, both work with or without a
context. But on Buildkite, the new "docker compose" only works for
certain contexts (like ECS local-simulation) but not for others (like
the default context).

Test Plan: unit

Reviewers: johann, dgibson

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D8690