[crag] Add mode to the pipeline runs filter
Summary: This diff adds mode to the PipelineRunsFilter object, adds a schema entry for mode in the runs table, and a method for data migration to retrieve the mode off of the pipeline run and add it to the corresponding row in the runs table.

Test Plan: Added backcompat tests for pg and sqlite, as well as tests to the run storage battery.

Reviewers: sandyryza, dish, alangenfeld, yuhan, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D8629