Solid -> Node
Summary:
Another in the rename stream, just changes the type name. Variables etc to be updated in future.

I think it'll be useful to have something like `Node[Solid]` & `Node[Graph]` or `SolidNode` and `SolidGraph` to distinguish, but for now just reflect the existing code operating in terms of a single type with getters to distinguish.

depends on D8714

Test Plan: bk

Reviewers: owen, yuhan, cdecarolis, prha

Reviewed By: owen

Differential Revision: https://dagster.phacility.com/D8715