docs: clarify how User Code images are created in Docker
Summary:
Resolves https://github.com/dagster-io/dagster/issues/3590.

We now explicitly mention what the Dockerfile should do - it should contain the
Dagster repository and any dependencies that it relies on.

However, in production scenarios, rather than copying user code and install dependencies individually,
the user is most likely to just install a python package containing all of their code. Here, we add
further clarification on that, mentioning that `dagsterApiGrpcArgs` can support this out of the box.

Test Plan: make dev

Reviewers: jordansanders, dgibson, catherinewu

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D8734