[dagit] Show start/end information on latest run in Job details
Summary:
On the job details header metadata, add a tooltip to the duration text to indicate the started/ended values of the run.

I don't really have enough space here to add these directly to the header instead of the duration, but hopefully this is useful.

Test Plan: View a job with a run, verify that the tooltip shows the correct information.

Reviewers: bengotow, sandyryza

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D8956