[memoization 8/n] make default fs io manager memoizable
Summary:
This diff makes the default fs io manager subclass MemoizableIOManager. This further decreases the friction users will encounter when they try to use memoization.

If this goes through, we can probably get rid of `versioned_fs_io_manager`. The question for me is whether we might want to preserve it for backcompat, or given the experimental-ness of all the memoization stuff, just force people to migrate.

Test Plan: Added additional unit test for memoization mode

Reviewers: sandyryza, alangenfeld

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D9085